Moving pc_data out of merc.h
